requirements.txt
absl-py==1.3.0
aiohttp==3.9.1
aiosignal==1.3.1
altair==4.2.0
anyio==3.7.1
appdirs==1.4.4
appnope==0.1.3
apptools==5.2.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astroid==2.11.7
asttokens==2.0.8
astunparse==1.6.3
async-timeout==4.0.3
attrs==22.1.0
autopep8==1.6.0
Babel==2.10.3
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
blinker==1.5
boto3==1.33.11
botocore==1.33.11
branca==0.5.0
CacheControl==0.13.1
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.0
colorama==0.4.5
commonmark==0.9.1
configobj==5.0.8
cycler==0.11.0
Cython==0.29.32
dacite==1.8.1
dask==2022.7.1
db-dtypes==1.0.4
deap==1.3.3
debugpy==1.6.3
decorator==5.1.1
DeepXDE==1.9.3
defusedxml==0.7.1
dicom-anonymizer==1.0.11
dill==0.3.6
distro==1.8.0
dm-tree==0.1.8
entrypoints==0.4
envisage==7.0.3
et-xmlfile==1.1.0
etils==0.8.0
exceptiongroup==1.2.0
executing==1.1.1
fastapi==0.104.1
fastjsonschema==2.16.2
filelock==3.13.1
flake8==4.0.1
flatbuffers==23.5.26
folium==0.12.1.post1
fonttools==4.38.0
frozenlist==1.4.1
fsspec==2022.10.0
gast==0.4.0
gensim==4.2.0
gitdb==4.0.9
GitPython==3.1.29
google-api-core==2.15.0
google-auth==2.23.0
google-auth-oauthlib==1.0.0
google-cloud-aiplatform==1.39.0
google-cloud-bigquery==3.14.1
google-cloud-bigquery-storage==2.16.2
google-cloud-core==2.3.2
google-cloud-documentai==2.21.1
google-cloud-resource-manager==1.6.3
google-cloud-storage==2.4.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.4.0
google-trans-new==1.1.9
googleapis-common-protos==1.56.4
grpc-google-iam-v1==0.13.0
grpcio==1.50.0
grpcio-status==1.48.2
h11==0.12.0
h5py==3.7.0
htmlmin==0.1.12
httpcore==0.15.0
httpx==0.23.0
hub-sdk==0.0.3
idna==3.4
ImageHash==4.3.1
imageio==2.22.2
imbalanced-learn==0.9.1
imblearn==0.0
importlib-metadata==5.0.0
importlib-resources==5.10.0
iniconfig==1.1.1
ipdb==0.13.9
ipykernel==6.15.3
ipympl==0.9.2
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==7.7.2
iso3166==2.1.1
isort==5.10.1
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.1.1
json5==0.9.10
jsonschema==4.16.0
jupyter-contrib-core==0.4.0
jupyter-contrib-nbextensions==0.5.1
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.5.0
jupyter-resource-usage==0.6.3
jupyter-server==1.21.0
jupyter_client==7.4.3
jupyter_core==4.11.2
jupyterlab==3.4.8
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.1
jupyterlab_server==2.16.1
kaggle==1.5.12
keras==2.13.1
keras-core==0.1.7
Keras-Preprocessing==1.1.2
keras-tuner==1.4.5
kiwisolver==1.4.4
kt-legacy==1.0.5
latexcodec==2.0.1
lazy-object-proxy==1.7.1
libclang==14.0.6
locket==1.0.0
lxml==4.9.1
mapbox==0.18.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
mayavi==4.8.1
mccabe==0.6.1
memoized-property==1.0.3
missingno==0.5.1
mistune==0.8.4
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
multimethod==1.8
namex==0.0.7
nbclassic==0.4.7
nbclient==0.7.0
nbconvert==6.5.4
nbformat==5.7.0
nbresult==0.0.9
nest-asyncio==1.5.6
networkx==2.8.7
nibabel==5.1.0
nltk==3.7
notebook==6.4.12
notebook-as-pdf==0.5.0
notebook_shim==0.2.0
numpy==1.23.4
oauthlib==3.2.2
openai==1.6.1
opencv-python==4.8.0.76
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.4
pandas-gbq==0.17.9
pandas-profiling==3.6.6
pandocfilters==1.5.0
parso==0.8.3
partd==1.3.0
patsy==0.5.3
pexpect==4.8.0
phik==0.12.2
pickleshare==0.7.5
Pillow==9.1.1
platformdirs==2.5.2
plotly==5.9.0
pluggy==1.0.0
pmdarima==2.0.1
polyline==2.0.1
prometheus-client==0.15.0
promise==2.3
prompt-toolkit==3.0.31
proto-plus==1.23.0
protobuf==3.20.3
psutil==5.9.3
psycopg2-binary==2.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
py-cpuinfo==9.0.0
pyaml==23.9.7
pyarrow==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybtex==0.24.0
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.9.2
pydata-google-auth==1.4.0
pydeck==0.8.0b4
pydicom==2.4.3
pyee==8.2.2
pyface==8.0.0
pyflakes==2.4.0
Pygments==2.13.0
pylint==2.14.5
Pympler==1.0.1
pyparsing==3.0.9
PyPDF2==3.0.1
pyppeteer==1.0.2
pyproject==1.3.1
PyQt6==6.5.3
PyQt6-Qt6==6.5.3
PyQt6-sip==13.6.0
pyrsistent==0.18.1
pytest==7.1.3
pytest-asyncio==0.19.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python-slugify==6.1.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq==24.0.1
regex==2022.9.13
requests==2.28.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
rich==12.6.0
rsa==4.9
s3transfer==0.8.2
SciANN==0.7.0.1
scikit-image==0.19.3
scikit-learn==1.3.2
scikit-optimize==0.9.0
scipy==1.9.3
seaborn==0.11.2
semver==2.13.0
Send2Trash==1.8.0
setuptools-scm==6.4.2
shapely==2.0.2
six==1.16.0
smart-open==6.2.0
smmap==5.0.0
sniffio==1.3.0
soundfile==0.12.1
soupsieve==2.3.2.post1
stack-data==0.5.1
starlette==0.27.0
statsmodels==0.13.2
stopit==1.1.2
streamlit==1.29.0
streamlit-datetime-range-picker==0.0.1
sympy==1.12
tangled-up-in-unicode==0.2.0
tenacity==8.1.0
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.13.0
tensorflow-datasets==4.6.0
tensorflow-decision-forests==1.5.0
tensorflow-estimator==2.13.0
tensorflow-macos==2.13.0
tensorflow-metadata==1.10.0
termcolor==2.0.1
terminado==0.16.0
text-unidecode==1.3
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tifffile==2022.10.10
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.5
toolz==0.12.0
torch==2.1.2
torchvision==0.16.2
tornado==6.2
TPOT==0.11.7
tqdm==4.64.1
traitlets==5.5.0
traits==6.4.2
traitsui==8.0.0
typeguard==2.13.3
typing_extensions==4.8.0
tzdata==2022.5
tzlocal==4.2
ultralytics==8.1.1
update-checker==0.18.0
uritemplate==4.1.1
urllib3==1.26.12
uvicorn==0.24.0.post1
validators==0.20.0
visions==0.7.5
vtk==9.2.6
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.1
websockets==10.4
Werkzeug==2.2.2
widgetsnbextension==3.6.1
wordcloud==1.9.2
wrapt==1.14.1
wurlitzer==3.0.3
xgboost==1.6.2
XlsxWriter==3.0.3
yapf==0.32.0
yarl==1.9.4
ydata-profiling==4.1.2
zipp==3.10.0
